// src/app/pages/infirmary/models/medicine.ts
export class Medicine {
    id?: number;
    ordenMedica: number;
    estado: string;
    fechaFormula: Date;
    medicamento: string;
    concentracion: string;
    dosis: string;
    viaAplicacion: string;
    goteo?: boolean;
    fRhora: string;
    horaAplicacion?:Date;
    unidadDeMedida?: string;
    cantidad: number;
    observacion: string;
    fechaAplicacion: Date;

}